Fix eslint module export warning
