+++ js/main.js
@@ -1,4 +1,53 @@
-import data from './data.js';
+const data = [
+  {
+    projectName: 'Tonic',
+    projectDescription:
+      'A daily selection of privately personalized reads; no accountsor sign-ups required.',
+    imageLink: 'images/Snapshoot-Portfolio1.png',
+    technologies: ['html', 'css', 'javascript'],
+    role: 'Back end Dev',
+    company: 'canopi',
+    year: '2015',
+    sourceLink: 'https://github.com/joseph07-drack/Portofolio/',
+    demoLink: 'https://joseph07-drack.github.io/Portofolio/',
+  },
+  {
+    projectName: 'Multi-post stories',
+    projectDescription:
+      'Experimental content creation feature that allows users to add to an existing story over the course of a day without spamming their friends.',
+    imageLink: 'images/Snapshoot-Portfolio2.png',
+    technologies: ['html', 'css', 'javascript'],
+    role: 'Full Stack Dev',
+    company: 'Facebook',
+    year: '2015',
+    sourceLink: 'https://github.com/joseph07-drack/Portofolio/',
+    demoLink: 'https://joseph07-drack.github.io/Portofolio/',
+  },
+  {
+    projectName: 'Facebook 360',
+    projectDescription:
+      "Exploring the future of media in Facebook's first Virtual Reality app; a place to discover and enjoy 360 photos and videos on Gear VR.",
+    imageLink: 'images/Snapshoot-Portfolio3.png',
+    technologies: ['html', 'Ruby on rails', 'css', 'javascript'],
+    role: 'Back End Dev',
+    company: 'Facebook',
+    year: '2015',
+    sourceLink: 'https://github.com/joseph07-drack/Portofolio/',
+    demoLink: 'https://joseph07-drack.github.io/Portofolio/',
+  },
+  {
+    projectName: 'Uber Navigation',
+    projectDescription:
+      'A smart assistant to make driving more safe, efficient, and fun by unlocking your most expensive computer: your car.',
+    imageLink: 'images/Snapshoot-Portfolio4.png',
+    technologies: ['html', 'Ruby on rails', 'css', 'javascript'],
+    role: 'Lead Developer',
+    company: 'Uber',
+    year: '2015',
+    sourceLink: 'https://github.com/joseph07-drack/Portofolio/',
+    demoLink: 'https://joseph07-drack.github.io/Portofolio/',
+  },
+];
 
 const mainMenu = document.querySelector('.mainMenu');
 const openMenu = document.querySelector('.openMenu');
